Develop

+++ app/Http/Controllers/TempleController.php
@@ -2,15 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\RoleAbility;
+use App\Http\Requests\PatchTempleMemberRolesRequest;
 use App\Http\Requests\StoreTempleRequest;
 use App\Http\Requests\UpdateTempleRequest;
 use App\Http\Resources\TempleResource;
 use App\Models\Temple;
-use Exception;
+use App\ValueObjects\Member;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use MongoDB\BSON\UTCDateTime;
-use PHPUnit\Event\Code\Throwable;
 
 class TempleController extends Controller
 {
@@ -36,6 +38,10 @@ class TempleController extends Controller
 
     public function destroy(Temple $temple)
     {
+        if (! Auth::user()->currentAccessToken()->canAny(RoleAbility::TEMPLE_DELETE->value, RoleAbility::TEMPLE_DELETE->value.':'.$temple->_id)) {
+            throw new AuthorizationException();
+        }
+
         return ['data' => ['success' => $temple->delete()]];
     }
 
@@ -46,25 +52,22 @@ class TempleController extends Controller
     {
         $user = Auth::user();
 
-        if (collect($temple->members)->where('username', Auth::user()->username)->count()) {
+        if (collect($temple->members)->where('username', $user->username)->first()) {
             return $this->error(trans('errors.temples.members.exists'));
         }
         
-        //  push user's temple
+        //  push to user's temple
         $user->increment('count_temples');
         $user->push('temples', [
             '_id' => $temple->id,
             'name' => $temple->name
         ], true);
 
-        //  push temple's member
+        //  push to temple's member
+        $user->roles = [];
+        $user->joined_at = new UTCDateTime();
+        $temple->push('members', (new Member($user->toArray()))->toArray());
         $temple->increment('count_members');
-        $temple->push('members', [
-            'username' => $user->username,
-            'role' => 'member',
-            'role_position' => 0,
-            'joined_at' => new UTCDateTIme()
-        ]);
 
         return ['data' => ['success' => true]];
     }
@@ -76,18 +79,29 @@ class TempleController extends Controller
     {
         $user = Auth::user();
 
-        if (! collect($temple->members)->where('username', Auth::user()->username)->count()) {
+        if (! collect($temple->members)->where('username', $user->username)->count()) {
             return $this->error(trans('errors.temples.members.not_exists'));
         }
         
-        //  push user's temple
+        //  pull from user's temple
         $user->decrement('count_temples');
         $user->pull('temples', ['_id' => $temple->id]);
 
-        //  push temple's member
+        //  pull from temple's member
         $temple->decrement('count_members');
         $temple->pull('members', ['username' => $user->username]);
 
         return ['data' => ['success' => true]];
+    }
+
+    public function patchMemberRoles(PatchTempleMemberRolesRequest $request, Temple $temple, $member)
+    {
+        if (! collect($temple->members)->where('username', $member)->count()) {
+            return $this->error(trans('errors.temples.members.not_exists'));
+        }
+
+        return ['data' => [
+            'success' => (boolean) $temple->where('members.username', 'mokilo')->update(['members.$.roles' => $request->roles])
+        ]];
     }
 }
